Unify logo into shared SiteLogo; /system uses article scroll-spin
Extract the blog/system logo into one SiteLogo component (signature vs
scroll-spin via a `spin` prop), replacing the one-off SystemLogo. /system
is a long-form scroll page, so it now spins like blog articles instead of
showing the static index signature.
